+++ test-shared-impl/src/main/java/tests/shared/inheritance/BasicDomainCommandType.java
@@ -1,0 +1,6 @@
+package tests.shared.inheritance;
+
+import io.domainlifecycles.domain.types.DomainCommand;
+
+public interface BasicDomainCommandType extends DomainCommand {
+}