+++ test-shared-impl/src/main/java/tests/shared/inheritance/ConcreteDomainCommand.java
@@ -1,0 +1,11 @@
+package tests.shared.inheritance;
+
+import lombok.Builder;
+
+public class ConcreteDomainCommand extends AbstractDomainCommand{
+
+    @Builder
+    public ConcreteDomainCommand(String msg) {
+        super(msg);
+    }
+}