+++ test-shared-impl/src/main/java/tests/shared/inheritance/ConcreteDomainEvent.java
@@ -1,0 +1,15 @@
+package tests.shared.inheritance;
+
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class ConcreteDomainEvent implements BasicDomainEventType{
+    final String msg;
+
+    @Builder
+    public ConcreteDomainEvent(String msg) {
+        this.msg = msg;
+    }
+}